[frontend | 고태현 | mod] 유지보수 용이를 위한 Component 분리
