+++ Route_Planning/src/App.tsx
@@ -1,4 +1,4 @@
-import RouteScreen from './RouteScreen'
+import RouteScreen from './components/RouteScreen'
 import './App.css'
 function App() {
